+++ HabboHotel/Rooms/RoomManager.cs
@@ -38,6 +38,7 @@ public class RoomManager : IRoomManager
     private readonly IRoomItemRemovalService _roomItemRemovalService;
     private readonly IRoomItemStateService _roomItemStateService;
     private readonly IRoomItemPlacementApplyService _roomItemPlacementApplyService;
+    private readonly IRoomItemTrackingService _roomItemTrackingService;
     private readonly IChatManager _chatManager;
     private readonly IBotManager _botManager;
     private readonly IRoomService _roomService;
@@ -69,6 +70,7 @@ public class RoomManager : IRoomManager
         IRoomItemRemovalService roomItemRemovalService,
         IRoomItemStateService roomItemStateService,
         IRoomItemPlacementApplyService roomItemPlacementApplyService,
+        IRoomItemTrackingService roomItemTrackingService,
         IGameClientManager gameClientManager,
         IGroupManager groupManager,
         IRoomService roomService,
@@ -98,6 +100,7 @@ public class RoomManager : IRoomManager
         _roomItemRemovalService = roomItemRemovalService;
         _roomItemStateService = roomItemStateService;
         _roomItemPlacementApplyService = roomItemPlacementApplyService;
+        _roomItemTrackingService = roomItemTrackingService;
         _clientManager = gameClientManager;
         _groupManager = groupManager;
         _roomService = roomService;
@@ -239,7 +242,7 @@ public class RoomManager : IRoomManager
 
     private Room CreateRoomInstance(RoomData data)
     {
-        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _roomItemLoadService, _roomItemRemovalService, _roomItemStateService, _roomItemPlacementApplyService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
+        return new Room(data, _clientManager, _database, _itemLoader, _roomItemPersistenceService, _roomItemPlacementValidatorService, _roomItemPlacementPersistenceService, _roomRollerService, _roomItemInventoryService, _roomItemUpdateQueueService, _roomItemLoadService, _roomItemRemovalService, _roomItemStateService, _roomItemPlacementApplyService, _roomItemTrackingService, _groupManager, _roomService, _chatManager, _botManager, _achievementService, _questService, _cacheManager, _languageManager, _itemTeleporterFinder, _itemHopperFinder, _badgeManager, _userDataFactory, this, _loggerFactory);
     }
 
     private static void DisposeRoom(Room room) => room.Dispose();